move type-guards to separate file
